Collected usage statistics of skims: 1. number of requests 2. number of times the actual value from skims map was returned Currently, collection is done for ODSkims, ParkingSkims and RideHailSkims
Resets stats at the beginning of each iteration, writes stats to logs at the end of each iteration. Writes stats to logs after generation of activities

+++ src/main/scala/beam/router/skim/core/AbstractSkimmer.scala
@@ -65,8 +65,8 @@ abstract class AbstractSkimmerReadOnly extends LazyLogging {
   private[core] var currentIterationInternal: Int = -1
   private[core] var aggregatedFromPastSkimsInternal = Map.empty[AbstractSkimmerKey, AbstractSkimmerInternal]
   private[core] val pastSkimsInternal = mutable.HashMap.empty[Int, Map[AbstractSkimmerKey, AbstractSkimmerInternal]]
-  var numberOfRequests: Int = 0
-  var numberOfSkimValueFound: Int = 0
+  var numberOfRequests: Long = 0
+  var numberOfSkimValueFound: Long = 0
   def currentIteration: Int = currentIterationInternal
   def aggregatedFromPastSkims: Map[AbstractSkimmerKey, AbstractSkimmerInternal] = aggregatedFromPastSkimsInternal
   def pastSkims: Map[Int, collection.Map[AbstractSkimmerKey, AbstractSkimmerInternal]] = pastSkimsInternal.toMap
